change some names, clean up code

+++ app/src/main/java/com/example/movieapp/data/datasource/local/entity/MoviesEntity.kt
@@ -5,7 +5,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "movies")
 data class MoviesEntity(
-    @PrimaryKey(autoGenerate = false) val id: Long,
+    @PrimaryKey(autoGenerate = false) val id: Int,
     val poster: String,
     val releaseDate: String,
     val rating: Double
